api/requirements.txt
google-generativeai==0.8.3
pydantic==2.9.2
pypdf2==3.0.1
pymupdf==1.25.2
pillow==11.0.0
